refactor: rename Client.tsx to descriptive component file names
